`BackupScreen` takes right sidebar as prop
